renamed minpermwer.py as min_perm_wer.py
